# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Neural Ram API
  description: |
    Neural Ram, persistent memory layer for AI agents.

    This document covers the HTTP/JSON REST surface. The following non-JSON
    transports are served by the same binary but intentionally not modeled
    here: the MCP Streamable HTTP endpoint at `/mcp`, the Server-Sent Events
    stream at `/v1/events`, Prometheus metrics at `/metrics`, and the HTML
    consent/landing shells served by `GET /authorize` and `GET /share/accept`.
    See docs/api.md for those.
  version: 0.4.0
  license:
    name: MIT
servers:
  - url: http://localhost:8674
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /v1/me/projects:
    get:
      operationId: meListProjects
      tags: [User]
      summary: List authenticated user's projects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProjects"

    post:
      operationId: meCreateProject
      tags: [User]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Slug conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/me/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: meGetProject
      tags: [User]
      summary: Get a project
      description: Retrieve a project by ID. Only the project owner can view their own projects.
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Not the project owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: meUpdateProject
      tags: [User]
      summary: Update a project
      description: Update a project's name, slug, description, default tags, or settings. Only the project owner can update their own projects.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
                default_tags:
                  type: array
                  items:
                    type: string
                settings:
                  type: object
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Not the project owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: meDeleteProject
      tags: [User]
      summary: Delete a project
      description: >-
        Permanently delete a project and all its memories, entities, relationships,
        and vectors. Only the project owner can delete their own projects. The
        'global' project cannot be deleted. Token usage records are preserved by
        reassigning them to the global project.
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Not the project owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT token or API key (prefixed with nram_k_)
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict, rate_limited, internal_error]
            message:
              type: string
            details:
              type: object

    # ── Core Models ────────────────────────────────────────────────────
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        namespace_id:
          type: string
          format: uuid
        owner_namespace_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        default_tags:
          type: array
          items:
            type: string
        settings:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ── Health ─────────────────────────────────────────────────────────
...
    PaginatedProjects:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ── User Self-Service ─────────────────────────────────────────────
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
          description: Auto-generated from name if omitted
        description:
          type: string
        default_tags:
          type: array
          items:
            type: string
...